requirements.txt
Django==3.2.8
django-allauth==0.44.0
django-braces==1.14.0
django-celery-results
django-cors-headers==3.6.0
django-dbbackup==3.3.0
django-extensions==3.1.0
django-filter==2.4.0
django-import-export==2.5.0
django-material-icons==0.1.2
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-oauth-toolkit==1.3.3
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-framework-social-oauth2==1.1.0
django-rest-swagger==2.2.0
django-storages==1.11.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
gunicorn==20.0.4
whitenoise==5.2.0
